pyproject.toml
[project]
name = "demo-dbos-pydantic-ai-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbos>=2.17.0",
    "pydantic-ai-slim[anthropic,openai]>=1.74.0",
]
